Permissions: detect burned permission manager
